Add support for LOGGED keyword, similar to UNLOGGED but for permanent
This extends the following statements with a new keyword called LOGGED,
to be able to force a relation to be permanent:
- CREATE SEQUENCE
- CREATE TABLE AS
- SELECT INTO
- CREATE TABLE

The implementation is done here with the introduction of a
RELPERSISTENCE_INVALID, which is set by the grammar when neither TEMP,
UNLOGGED or LOGGED are specified.  This is handy for an upcoming patch
that aims to introduce relpersistence inheritance for partitions, based
on its partitioned table.