Extend custom cumulative stats to be persistent
This patch uses the approach of a unique file, so as all the stats are
stored in the existing PGSTAT_STAT_PERMANENT_FILENAME, extending its
format with a new entry type to be able to read on load and write at
shutdown the custom kinds of statistics registered:
- The name of the custom stats kinds is stored into the stats file
rather than its ID.
- The rest of the hash key, made of the database OID and the object OID,
is stored if the custom kind is not serialized.  If serialized, where
callbacks are used to do a mapping of the on-disk name <-> hash key, the
serialized name is pushed instead.

The patch includes an example of what can be achieved with the test
module injection_points, with a TAP test checking if statistics can be
kept after a clean restart.