injection_points: Add basic isolation test
This can act as a template when implementing an isolation test with
injection points, and tracks in a much simpler way some of the behaviors
implied in the existing isolation test "inplace".  Particularly, a
detach should not affect a backend wait; a wait can should only be
stopped by a wakeup.