Add the pg_stat_get_backend_stats() helper for pg_stat_get_backend_io()
This commit adds pg_stat_get_backend_stats(), a helper routine for
pg_stat_get_backend_io(), that returns the backend stats based on a pid passed
as an argument.

This will be used in a follow-up commit that uses the same logic to return the
per backend WAL stats.