Fix possible double-release of spinlock in procsignal.c
9d9b9d46f3c5 has added spinlocks to protect the fields in ProcSignal
flags, introducing a code path in ProcSignalInit() where a spinlock
could be released twice if the pss_pid field of a ProcSignalSlot is set.

The sanity check generates a LOG that can be delayed until after the
spinlock is released as, like older versions up to v17, the code expects
the initialization of the ProcSignalSlot to happen even if pss_pid is
found incorrect.  The code is changed so as the old pss_pid is read
while holding the slot's spinlock, with the LOG generated after
releasing the spinlock, preventing any double release.

Discussion: https://postgr.es/m/dca47527-2d8b-4e3b-b5a0-e2deb73371a4@postgrespro.ru