Fix some gaps with pg_stat_io and WAL-related processes
The WAL receiver and WAL summarizer processes gain each one a call to
pgstat_report_wal(), to make sure that they report their WAL statistics
in pg_stat_io.

pg_stat_io is adjusted so as these processes do not report any rows when
the IOObject is not WAL, making the view easier to use with less rows.