Fix Normalization for squashed query texts
62d712ec added the ability to squash constants from an
IN list/ArrayExpr for queryId computation purposes. However,
in certain cases, this broke normalization. For example,
"IN (1, 2, int4(1))" is normalized to "IN ($2 /*, ... */))",
which leaves an extra parenthesis at the end of the normalized string.

To correct this, the start and end boundaries of an expr_list are
now tracked by the various nodes used during parsing and are made
available to the ArrayExpr node for query jumbling. Having these
boundaries allows normalization to precisely identify the locations
in the query text that should be squashed.