Fix broken normalization due to duplicate constant locations
pg_stat_statements anticipates that certain constant
locations may be recorded multiple times and attempts
to avoid calculating a length for these locations in
fill_in_constant_lengths.

However, during generate_normalized_query, these
locations are not excluded from consideration and
will increment the $n counter for every recorded
occurrence of such a location. In practice, this can
lead to incorrect normalization in certain cases.

select where '1' IN ('2'::int, '3'::int::text)

would be normalized to:

select where $1 IN ($3, $4)

instead of the correct:

select where $1 IN ($2, $3)

This is because the left-expression, '1' is used
as an argument in the OpExpr generated for every
element in the IN clause.

To correct, track the number of constants replaced
with an $n by a separate counter instead of the
iterator used to loop through the list of locations.