Abort connection when sync is sent after a COPY
When the backend reads COPY data, it ignores all Sync messages. With
psql pipelining, it's possible to manually send Sync with \sendpipeline
which leaves the frontend in an unrecoverable state as the backend won't
send the necessary ReadyForQuery.

It could be possible to artificially reduce the piped_syncs and
requested_results, but libpq's state will still have queued Sync in its
command queue. And the only way to consume those without directly
calling pqCommandQueueAdvance is to process ReadyForQuery messages that
won't be sent since the backend ignored the Syncs.

Thus, this patch abort the connection if we detect excessive sync after
a COPY in a pipeline to avoid staying in an inconsistent protocol state.