Introduce global 64-bit TOAST ID counter
This is added to the control file, providing a unique 64-bit-wide source
for toast value IDs, with the same guarantees as OIDs in terms of
durability.  SQL functions and tools looking at the control file are
updated.  A WAL record is generated every 8k values generated, that can
be adjusted if required.