Forbid usage of COPY in a pipeline
Running COPY within a pipeline can break protocol synchronisation in multiple
ways.
- In copy mode, the backend process ignores additional Sync messages and
won't send the matching ReadyForQuery expected by the frontend. Doing a
\syncpipeline just after COPY will leave the frontend waiting for a
ReadyForQuery message that won't be sent.
- libq automatically sends a Sync with the Copy message which isn't
tracked in the command queue, creating an unexpected synchronisation point. A
failure during copy will leave the pipeline in an aborted state while the
backend would be in a clean state, ready to process commands.

Fixing those issues would require modifications in how libpq handle
pipeline and COPY. Since this is a new feature and the use case of COPY
in a pipeline has limited usefulness, it was decided to forbid the use
of COPY within a pipeline to avoid possible break of protocol
synchronisation.