Introduce global 64-bit TOAST ID counter in control file
An 8 byte counter is added to the control file, providing a unique
64-bit-wide source for toast value IDs, with the same guarantees as OIDs
in terms of durability.  SQL functions and tools looking at the control
file are updated.  A WAL record is generated every 8k values generated,
that can be adjusted if required.

Requires a bump of WAL format.
Requires a bump of control file version.
Requires a catalog version bump.