Refactor some TOAST value ID code to use uint64 instead of Oid
This change is a mechanical switch to change most of the code paths that
assume TOAST value IDs to be Oids to become uint64, easing an upcoming
change to allow 8-byte TOAST values.

The areas touched are related to table AM, amcheck and logical
decoding's reorder buffer.  A good chunk of the changes involve
switching printf() markers from %u to PRIu64.