Add support for TOAST table types in pg_dump and pg_restore
This includes the possibility to perform binary upgrades with TOAST
table types applied to a new cluster, relying on SET commands based on
default_toast_type to apply one type of TOAST table or the other.

Some tests are included, this is a pretty mechanical change.

Dump format is bumped to 1.17 due to the addition of the TOAST table
type in the custom format.