Add support for bigint TOAST values
This commit adds the possibility to define TOAST tables with bigint as
value ID, baesd on the GUC default_toast_type.  All the external TOAST
pointers still rely on varatt_external and a single vartag, with all the
values inserted in the bigint TOAST tables fed from the existing OID
value generator.  This will be changed in an upcoming patch that adds
more vartag_external types and its associated structures, with the code
being able to use a different external TOAST pointer depending on the
attribute type of chunk_id in TOAST relations.

All the changes done here are mechanical, with all the TOAST code able
to do chunk ID lookups based on the two types supported.

XXX: Catalog version bump required.