Refactor external TOAST pointer code for better pluggability
This commit introduces a new interface for external TOAST pointers,
which is able to make a translation of the varlena pointers stored on
disk to/from an new in-memory structure called toast_external.  The
types of varatt_external supported on disk need to be registered into a
new subsystem in a new file, called toast_external.[c|h], then define a
set of callbacks to allow the toasting and detoasting code to use it.

The existing varatt_external is renamed to varatt_external_oid, to map
with the fact that this structure is used to store external TOAST
pointers based on OID values.

A follow-up change will rely on this refactoring to introduce a new type
of vartag_external with an associated varatt_external that is able to
store 8-byte value IDs on disk.