Fix performance regression with pgstats reports
The introduction of the new callback to control the flush of pending
entries is proving to show up high in workloads that do not report any
stats (read only, no function calls, etc).

This commit switches the code to use a cheaper approach, with a single
boolean flag that can be switched to "true" by any fixed-numbered stats
kinds to force pgstat_report_stat() to go through one round of reports.
The flag can only be reset by pgstat_report_stat() once a full round of
reports is done.

The callback have_static_pending_cb can be removed as an effect of that.