Improve test coverage of date_trunc() on infinity
Commit d85ce012f99 added new error handling code to date_trunc() of
timestamp, timestamptz, and interval with infinite values.  But the
new test cases added by that commit did not actually test the new
code, since they used the "not recognized" unit 'ago' but not a "not
supported" unit, which is what the new code was dealing with.  This
patch adds some test cases to cover the new code.  I'm using
'timezone' as the not supported unit.