Fix incorrect Datum conversion in timestamptz_trunc_internal()
It used PG_RETURN_TIMESTAMPTZ(), but the return type is TimestampTz,
not Datum.  On 64-bit systems, there is no effect, since this just
ends up casting 64-bit integers back and forth.  But on 32-bit
systems, timestamptz is pass-by-reference, and so
PG_RETURN_TIMESTAMPTZ() allocates new memory and returns the address,
but the caller will try to interpret this as a timestamp value.  The
effect is that date_trunc(..., 'infinity'::timestamptz) will return
random values (instead of the correct return value 'infinity').

The fix is to use a straight "return" call.

Commit d85ce012f99 added this code but provided tests that did not
cover this correctly.  This was fixed in the previous patch.

Reviewed-by: Tom Lane <tgl@sss.pgh.pa.us>
Discussion: https://www.postgresql.org/message-id/flat/8246d7ff-f4b7-4363-913e-827dadfeb145%40eisentraut.org