Move static inline routines of varatt_external_oid to toast_external.c
This isolates most of the knowledge of varatt_external_oid into the
local area where it is manipulated through the toast_external transition
type, with the backend code not requiring it.  Extension code should not
need it either, as toast_external should be the layer to use when
looking at external on-dist TOAST varlenas.