Introduce new callback to get fresh TOAST values
This callback is called by toast_save_datum() to retrieve a new value
from a source related to the vartag_external we are dealing with.  As of
now, it is simply a wrapper around GetNewOidWithIndex() for the "OID"
on-disk TOAST external pointer.

This will be used later on by more external pointer types, like the int8
one.

InvalidToastId is introduced to track the concept of an "invalid" TOAST
value, required for toast_save_datum().