Enlarge OID generation to 8 bytes
This adds a new routine called GetNewObjectId8() in varsup.c, which is
able to retrieve a 64b OID.  GetNewObjectId() is kept compatible with
its origin, where we still check that the lower 32 bits of the counter
do not wraparound, handling the FirstNormalObjectId case.

pg_resetwal -o/--next-oid is updated to be able to handle 8-byte OIDs.