Add support for entry counting in pgstats
Stats kinds can set an option call track_counts, that will make pgstats
track the number of entries that exist in the shared hashtable.

As there is only one code path where a new entry is added, and one code
path where entries are removed, the count tracking is rather
straight-forward.  Reads of these counters are optimistic, and may
change across two calls.

A use case of this facility is PGSS, where we need to be able to cap the
number of entries that would be stored in the shared hashtable, based on
its "max" GUC.