Exit copy state in pgbench to avoid infinite loop
Currently, pgbench aborts when a copy response is received in
readCommandResponse. However, PQgetResult will return an empty
result when there's no async result through getCopyResult, leading to an
infinite loop in the error handling.

This patch forcefully exits the copy state with PQendcopy before moving
to the error handler, avoiding the infinite loop.