Fix orphaned origin in shared memory after subscription drop
Currently, we store replication origin information in both the catalog and
shared memory (once the origin is set up). During a transaction, if a
replication origin is created and then set up, but the transaction subsequently
aborts, only the catalog changes are rolled back, leaving an orphaned origin in
shared memory. This is confusing because users can still see orphaned records
in pg_replication_origin_status even after dropping the subscription, and this
prevents other origins from reusing the position.

This issue is more likely to happen during table synchronization, where the
replication origin is created within the same transaction that performs the
initial copy. If the COPY operation fails, the origin in shared memory remains
uncleared.

To address this, the commit moves origin creation to the end of a separate
transaction preceding the one executing the initial COPY. This prevents the
risk that changes to the origin in shared memory cannot be rolled back if the
transaction aborts.